harden PAO and SRIOV config util

+++ PAO/cleanup.sh
@@ -1,13 +1,16 @@
 #!/bin/sh
 #
+# Cleanup performanceprofile while taking into consideration the SRIOV states.
+# 	If SRIOV is still active, leave node labels and MCP alone.
+#
 
 set -euo pipefail
 
 source ./setting.env
 source ./functions.sh
+SINGLE_STEP=${SINGLE_STEP:-false}
 
 parse_args $@
-
 
 ##### Remove performan-profile ######
 echo "Removing performance profile ..."
@@ -22,22 +25,25 @@ else
   echo "No performance profile: done"
 fi
 
-echo HN short; exit
+if oc get SriovNetworkNodePolicy &>/dev/null; then
+    echo "SRIOV is still active. Skip the rest. Done"
+    exit
+fi
+
+
+echo "Next, remove node labels and MCP ${MCP}"
+prompt_continue 
 
 echo "deleting label for $WORKER_LIST ..."
 for worker in $WORKER_LIST; do
-   oc label --overwrite node ${worker} node-role.kubernetes.io/${MCP}-
+    oc label --overwrite node ${worker} node-role.kubernetes.io/${MCP}-
 done
-
-# wait for nodes to move back to mcp worker
-if [[ "${WAIT_MCP}" == "true" ]]; then
-    wait_mcp worker
-fi
 
 ##### Remove MCP ######
 if oc get mcp $MCP 2>/dev/null; then
     oc delete -f ${MANIFEST_DIR}/mcp-${MCP}.yaml
     echo "deleted mcp for ${MCP}: done"
 fi
+
 
 # EOF
